+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 * New syntax, attribute and declaration kinds introduced in Swift 6.1-6.3.  
   [John Fairhurst](https://github.com/johnfairh)
+* Improve reporting of `sourcekitdInProc` loading failures.  
+  [Daniel Sunarjo](https://github.com/sunarjodaniel)
 
 #### Bug Fixes